services/auth-service/requirements.txt
Flask==2.1.2
Flask-JWT-Extended==4.3.0
Flask-SQLAlchemy==2.5.1
requests==2.26.0
python-dotenv==0.19.2